Implementing BaseActivity to deal with OptionsMenu in both activities.
