+++ app/src/main/java/com/support/android/designlibdemo/CheeseDetailActivity.java
@@ -19,17 +19,12 @@ package com.support.android.designlibdemo;
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.design.widget.CollapsingToolbarLayout;
-import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
-import android.view.Menu;
-import android.view.MenuItem;
 import android.widget.ImageView;
 
 import com.bumptech.glide.Glide;
 
-import java.util.Random;
-
-public class CheeseDetailActivity extends AppCompatActivity {
+public class CheeseDetailActivity extends BaseActivity {
 
     public static final String EXTRA_NAME = "cheese_name";
 
@@ -55,11 +50,5 @@ public class CheeseDetailActivity extends AppCompatActivity {
     private void loadBackdrop() {
         final ImageView imageView = (ImageView) findViewById(R.id.backdrop);
         Glide.with(this).load(Cheeses.getRandomCheeseDrawable()).centerCrop().into(imageView);
-    }
-
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        getMenuInflater().inflate(R.menu.sample_actions, menu);
-        return true;
     }
 }
